requirements.txt
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
scikit-image>=0.22
scipy>=1.11
pyyaml>=6.0
tqdm>=4.65
torch>=2.1
torchvision>=0.16
transformers>=4.40
timm>=1.0.8
open-clip-torch>=2.24.0
opencv-python>=4.9
git+https://github.com/isl-org/MiDaS.git
kornia>=0.7.1
Pillow>=10.0
seaborn>=0.13
imageio>=2.34
pytest>=8.0
notebook>=7.1
jupyterlab>=4.2
wandb>=0.17.0
hydra-core>=1.3
